Activity enrichment now ignores missing schemas
Before, when the activity enricher found a token that doesn't have
a matching schema, a MissingSchemaError error would be thrown.

This commit changes the enrichment process to ignore values for
which a matching schema can't be found. This will allow us to pass
activities to the enricher without having to worry about values that
don't need to be enriched in the first place.